feat(shared-modules-refactor): Sub-PR 4 — extraer <AdminJuntasListModule>
Sub-PR 4 de la iniciativa shared-modules-refactor. Auditoría previa
concluyó que del triángulo `inicio/juntas` ↔ `rdb/admin/juntas` ↔
`dilesa/admin/juntas` solo RDB+DILESA admin son el mismo módulo —
`/inicio/juntas` es vista personal multi-empresa standalone.

## Decisión técnica

Adoptamos DILESA como base correcta. RDB hereda todas las features:

- Auto-title generator (`"YYYY-MM-DD, HH:MM AM/PM - Tipo"`).
- Filtro por mes.
- Columna `JuntaContentPreview` (TEMP, deuda de migración Coda).
- Task counts granulares (Avanzadas/Terminadas).
- Búsqueda en título + descripción (plain text).
- Sheet UX con flujo "Iniciar Junta" en lugar de Dialog modal.

Sin feature flags por empresa — la convención SM3 es "todas las empresas
tienen el mismo comportamiento, salvo cosmética".

## Cambios

Componente nuevo:
- `components/juntas/admin-juntas-list-module.tsx` — todo el flow de
  listado admin parametrizado por `empresaId`, `empresaSlug`, `title`.

Pages reducidos:
- `app/rdb/admin/juntas/page.tsx`: 502 → 14 líneas.
- `app/dilesa/admin/juntas/page.tsx`: 687 → 14 líneas.

Bug oportunístico arreglado:
- `app/inicio/juntas/page.tsx` y `app/inicio/juntas/[id]/page.tsx`
  tenían `<RequireAccess empresa="rdb">` hardcoded — bloqueaba acceso
  a usuarios DILESA al dashboard personal de juntas. Resuelto quitando
  el wrap `<RequireAccess>` (mismo patrón que `/inicio/tasks`: el proxy
  ya valida sesión).

## Validación

- typecheck ✅
- lint ✅ (4 warnings pre-existentes en `/inicio/juntas/[id]/page.tsx`,
  no introducidos por mis cambios)
- format:check ✅
- 222 tests ✅

Smoke pendiente en preview deploy:
- /rdb/admin/juntas: lista carga, filtros (estado/tipo/mes),
  búsqueda, crear junta con auto-title, content preview, task counts.
- /dilesa/admin/juntas: idem.
- /inicio/juntas: usuarios DILESA pueden abrir el dashboard personal
  (verificar que ya no bloquea con "Acceso restringido").
