+++ app/inicio/juntas/[id]/page.tsx
@@ -5,7 +5,6 @@
  * data. Proper typing requires schema refactor — out of scope for lint cleanup.
  */
 
-import { RequireAccess } from '@/components/require-access';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 import { createSupabaseERPClient } from '@/lib/supabase-browser';
@@ -1847,10 +1846,10 @@ function JuntaDetailInner() {
   );
 }
 
+// Sin RequireAccess de empresa: el dashboard personal es para todo usuario
+// logueado; el proxy ya valida la sesión. Mismo patrón que `/inicio/tasks`.
+// (Antes había `<RequireAccess empresa="rdb">` hardcoded — bloqueaba acceso
+// a usuarios DILESA. Bug arreglado en Sub-PR 4 de `shared-modules-refactor`.)
 export default function Page() {
-  return (
-    <RequireAccess empresa="rdb">
-      <JuntaDetailInner />
-    </RequireAccess>
-  );
+  return <JuntaDetailInner />;
 }
